feat: create Navigation with Links
